Arch/Dev: Merge StlExtensions with IOHelpers, move into lib/
