+++ src/core/r4300/gameshark.cpp
@@ -5,7 +5,6 @@
  */
 
 #include "stdafx.h"
-#include <StlExtensions.h>
 #include <core/Core.h>
 #include <core/memory/memory.h>
 #include <core/r4300/gameshark.h>
