+++ src/stdafx.h
@@ -41,3 +41,4 @@
 #include <deque>
 #include <numeric>
 #include <spdlog/logger.h>
+#include <IOHelpers.h>
