+++ src/view/gui/Main.h
@@ -31,7 +31,7 @@
 #endif
 
 #define BASE_NAME L"Mupen 64 "
-#define CURRENT_VERSION L"1.1.9-9"
+#define CURRENT_VERSION L"1.2.0-split-1"
 
 #define MUPEN_VERSION BASE_NAME CURRENT_VERSION ARCH_INFO CHARSET_INFO BUILD_TARGET_INFO
