+++ src/view/gui/features/ConfigDialog.cpp
@@ -9,7 +9,6 @@
 #include <Config.h>
 #include <FrontendService.h>
 #include <Messenger.h>
-#include <StlExtensions.h>
 #include <Windows.h>
 #include <core_api.h>
 #include <resource.h>
